@@ -1,13 +1,13 @@
 {
   "name": "alerthub",
-  "version": "2.2.2",
+  "version": "2.2.3",
   "description": "A tool to send notifications when an update to a repository is released",
-  "main": "./src/index.mjs",
+  "main": "./src/index.js",
   "type": "commonjs",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "init": "cp ./etc/config.example.js ./etc/config.js",
-    "start": "node ./src/index.mjs"
+    "start": "node ./src/index.js"
   },
   "author": "Arda Kilicdagi",
   "license": "MIT",
@@ -16,19 +16,18 @@
     "url": "https://github.com/Ardakilic/alerthub.git"
   },
   "dependencies": {
-    "nodemailer": "^6.7.8",
+    "node-telegram-bot-api": "^0.66.0",
+    "nodemailer": "^6.9.15",
     "pushbullet": "^3.0.0",
+    "pushover-notifications": "^1.2.3",
     "rss-braider": "^1.2.4",
     "rss-feed-emitter": "^3.2.3",
-    "striptags": "^3.2.0",
-    "pushover-notifications": "^1.2.2"
-  },
-  "devDependencies": {
-    "eslint": "^8.22.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.26.0"
+    "striptags": "^3.2.0"
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=20.0.0"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.9.0"
   }
 }
