@@ -1,10 +1,10 @@
 import mail from 'nodemailer';
 import striptags from 'striptags';
 
-export default class Email {
+export default {
   // Sends and e-mail notification to provided user
-  static async sendEmailNotification(config, feedData) {
-    const transporter = mail.createTransport(config.notifications.email.config);
+  async sendEmailNotification(config, feedData) {
+    const transporter = mail.createTransport(config.config);
 
     /* await smtp.verify((error, success) => {
       if (error) {
@@ -15,9 +15,9 @@ export default class Email {
     }); */
     // E-mail options that will be used by nodemailer
     const mailOptions = {
-      from: config.notifications.email.mailOptions.from,
-      to: config.notifications.email.mailOptions.to,
-      subject: config.notifications.email.mailOptions.subjectPrefix.length > 0 ? `${config.notifications.email.mailOptions.subjectPrefix} - ${feedData.title}` : feedData.title,
+      from: config.mailOptions.from,
+      to: config.mailOptions.to,
+      subject: config.mailOptions.subjectPrefix.length > 0 ? `${config.mailOptions.subjectPrefix} - ${feedData.title}` : feedData.title,
       text: `${feedData.link}\n${striptags(feedData.description)}`,
       html: `<a target="_blank" href="${feedData.link}">${feedData.title}</a><br><br>${feedData.description}`,
     };
@@ -29,7 +29,8 @@ export default class Email {
         return false;
       }
       // console.log('Message sent: %s', info.messageId);
+      // return Promise.resolve(info);
       return info;
     });
-  }
+  },
 }
